Optimize and Rename "Logg class"
