+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,3 @@
-#Sat Jan 19 00:05:04 KST 2019
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
